# geoai-service/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoAI Tool Adapter Service",
    "description": "Controlled, typed tool surface between an AI agent and the GeoSphere platform's PostGIS database, MinIO object storage, and internal APIs. The agent never receives credentials or internal URLs \u2014 every call here is validated, rate-limited, cached, and logged.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/geoai/geocode/reverse": {
      "post": {
        "tags": [
          "geosphere-adapters"
        ],
        "summary": "Reverse Geocode",
        "operationId": "reverse_geocode_geoai_geocode_reverse_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReverseGeocodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReverseGeocodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/geoai/geocode/search": {
      "post": {
        "tags": [
          "geosphere-adapters"
        ],
        "summary": "Search Place",
        "operationId": "search_place_geoai_geocode_search_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchPlaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPlaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/geoai/route": {
      "post": {
        "tags": [
          "geosphere-adapters"
        ],
        "summary": "Get Route",
        "operationId": "get_route_geoai_route_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/geoai/land-record": {
      "post": {
        "tags": [
          "geosphere-adapters"
        ],
        "summary": "Land Record",
        "operationId": "land_record_geoai_land_record_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LandRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/geoai/environment": {
      "post": {
        "tags": [
          "geosphere-adapters"
        ],
        "summary": "Environment Snapshot",
        "operationId": "environment_snapshot_geoai_environment_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentSnapshotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentSnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/geoai/dataset-layer": {
      "post": {
        "tags": [
          "geosphere-adapters"
        ],
        "summary": "Dataset Layer",
        "operationId": "dataset_layer_geoai_dataset_layer_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetLayerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetLayerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DatasetLayerRequest": {
        "properties": {
          "layer": {
            "type": "string",
            "title": "Layer"
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "layer"
        ],
        "title": "DatasetLayerRequest"
      },
      "DatasetLayerResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "default": "success"
          },
          "feature_collection": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Collection"
          }
        },
        "type": "object",
        "title": "DatasetLayerResponse"
      },
      "EnvironmentSnapshotRequest": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "EnvironmentSnapshotRequest"
      },
      "EnvironmentSnapshotResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "default": "success"
          },
          "weather": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather"
          },
          "air_quality": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Air Quality"
          }
        },
        "type": "object",
        "title": "EnvironmentSnapshotResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LandRecordRequest": {
        "properties": {
          "district": {
            "type": "string",
            "title": "District"
          },
          "taluk": {
            "type": "string",
            "title": "Taluk"
          },
          "hobli": {
            "type": "string",
            "title": "Hobli"
          },
          "village": {
            "type": "string",
            "title": "Village"
          },
          "survey": {
            "type": "string",
            "title": "Survey"
          },
          "surnoc": {
            "type": "string",
            "title": "Surnoc",
            "default": "*"
          },
          "hissa": {
            "type": "string",
            "title": "Hissa",
            "default": "*"
          }
        },
        "type": "object",
        "required": [
          "district",
          "taluk",
          "hobli",
          "village",
          "survey"
        ],
        "title": "LandRecordRequest"
      },
      "LandRecordResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "default": "success"
          },
          "owners": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Owners"
          },
          "use_case": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Case"
          }
        },
        "type": "object",
        "title": "LandRecordResponse"
      },
      "LatLon": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "LatLon"
      },
...
      "PlaceResult": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          }
        },
        "type": "object",
        "required": [
          "label",
          "lat",
          "lon"
        ],
        "title": "PlaceResult"
      },
      "ReverseGeocodeRequest": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "ReverseGeocodeRequest"
      },
      "ReverseGeocodeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "default": "success"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "place_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Name"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          }
        },
        "type": "object",
        "title": "ReverseGeocodeResponse"
      },
      "RouteRequest": {
        "properties": {
          "origin": {
            "$ref": "#/components/schemas/LatLon"
          },
          "destination": {
            "$ref": "#/components/schemas/LatLon"
          },
          "mode": {
            "type": "string",
            "enum": [
              "driving",
              "walking",
              "cycling"
            ],
            "title": "Mode",
            "default": "driving"
          }
        },
        "type": "object",
        "required": [
          "origin",
          "destination"
        ],
        "title": "RouteRequest"
      },
      "RouteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "default": "success"
          },
          "distance_meters": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Meters"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry",
            "description": "GeoJSON LineString"
          }
        },
        "type": "object",
        "title": "RouteResponse"
      },
      "SearchPlaceRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchPlaceRequest"
      },
      "SearchPlaceResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "default": "success"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PlaceResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "SearchPlaceResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}